java/src/main/java/com/studio3104/leetcode/explore/collection/medium/TreesAndGraphs/BinaryTreeInorderTraversal/Solution.java
package com.studio3104.leetcode.explore.collection.medium.TreesAndGraphs.BinaryTreeInorderTraversal;

import com.studio3104.leetcode.structure.TreeNode;

import java.util.ArrayList;
import java.util.List;

class Solution {
    private void inorderTraversal(TreeNode node, List<Integer> result) {
        if (node == null) {
            return;
        }

        inorderTraversal(node.left, result);
        result.add(node.val);
        inorderTraversal(node.right, result);
    }

    public List<Integer> inorderTraversal(TreeNode root) {
        List<Integer> result = new ArrayList<>();
        inorderTraversal(root, result);
        return result;
    }
}
